# tribute-deploy/tribute-deploy/api-client-src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Pastor Joseph Olusola Samuel @ 60
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Tribute {
  id: number;
  authorName: string;
  relationship: string;
  message: string;
  createdAt: string;
}

export interface TributeInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  authorName: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  relationship: string;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  message: string;
}

export interface TributeStats {
  totalCount: number;
  recentTributes: Tribute[];
}

export type RsvpAttendance = typeof RsvpAttendance[keyof typeof RsvpAttendance];


export const RsvpAttendance = {
  'Yes,_I_will_attend_in_person': 'Yes, I will attend in person',
  I_will_attend_virtually: 'I will attend virtually',
  'No,_I_will_not_be_able_to_attend': 'No, I will not be able to attend',
} as const;

export interface Rsvp {
  id: number;
  fullName: string;
  phone: string;
  attendance: RsvpAttendance;
  createdAt: string;
}

export type RsvpInputAttendance = typeof RsvpInputAttendance[keyof typeof RsvpInputAttendance];


export const RsvpInputAttendance = {
  'Yes,_I_will_attend_in_person': 'Yes, I will attend in person',
  I_will_attend_virtually: 'I will attend virtually',
  'No,_I_will_not_be_able_to_attend': 'No, I will not be able to attend',
} as const;

export interface RsvpInput {
  /**
     * @minLength 1
     * @maxLength 150
     */
  fullName: string;
  /**
     * @minLength 1
     * @maxLength 30
     */
  phone: string;
  attendance: RsvpInputAttendance;
}

export interface RsvpStats {
  total: number;
  inPerson: number;
  virtual: number;
  notAttending: number;
}

export interface ApiError {
  error: string;
}